Add block support to world seed load csv method, and guard clause for model deletion

+++ db/world_seed.rb
@@ -1,21 +1,23 @@
 require 'csv'
 
 class WorldSeed
-  def self.load_csv(filename, model_class)
+  def self.load_csv(filename, model_class, &block)
     raise "Don't run this in production!" if Rails.env.production?
 
-    model_class.delete_all
+    model_class.delete_all if model_class.table_exists?
     models = []
     path = Rails.root.join('db', 'seeds', filename)
-    CSV.foreach(path, headers: true) do |row|
-      models << model_class.new(row.to_h)
+    if block.nil?
+      CSV.foreach(path, headers: true) do |row|
+        models << model_class.new(row.to_h)
+      end
+    else
+      CSV.foreach(path, headers: true) do |row|
+        models << block.call(row)
+      end
     end
     model_class.import models, recursive: true
   end
-
-  # def self.verify (model, filename)
-  #
-  # end
 
   private
